+++ ouroboros-network/src/Ouroboros/Network/PeerSharing.hs
@@ -9,22 +9,32 @@ module Ouroboros.Network.PeerSharing
   , peerSharingClient
   , peerSharingServer
   , requestPeers
+  , PeerSharingAPI (..)
+  , newPeerSharingAPI
+    -- * Constants
+  , ps_POLICY_PEER_SHARE_STICKY_TIME
+  , ps_POLICY_PEER_SHARE_MAX_PEERS
   ) where
 
 import Control.Applicative (Alternative)
 import Control.Concurrent.Class.MonadMVar (MVar, MonadMVar (putMVar),
            newEmptyMVar, takeMVar)
-import Control.Concurrent.Class.MonadSTM.Strict (MonadSTM, STM, StrictTMVar,
-           StrictTVar, atomically, modifyTVar, newEmptyTMVarIO, newTVarIO,
-           putTMVar, retry, takeTMVar)
+import Control.Concurrent.Class.MonadSTM.Strict
 import Control.Monad.Class.MonadThrow (MonadThrow, bracket)
+import Control.Monad.Class.MonadTime.SI
+import Data.Hashable (Hashable (..))
+import Data.List (sortBy)
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
 import Data.Monoid.Synchronisation (FirstToFinish (..), runFirstToFinish)
+import Data.Set qualified as Set
 import Ouroboros.Network.ControlMessage (ControlMessage (..), ControlMessageSTM)
+import Ouroboros.Network.PeerSelection.Governor.Types (PublicPeerSelectionState,
+           availableToShare, emptyPublicPeerSelectionState)
 import Ouroboros.Network.Protocol.PeerSharing.Client (PeerSharingClient (..))
 import Ouroboros.Network.Protocol.PeerSharing.Server (PeerSharingServer (..))
 import Ouroboros.Network.Protocol.PeerSharing.Type (PeerSharingAmount)
+import System.Random
 
 -- | Request and Result queue for the peer sharing client implementation.
 --
@@ -104,11 +114,104 @@ peerSharingClient controlMessageSTM
         putMVar resultQueue result
         peerSharingClient controlMessageSTM psc
 
-peerSharingServer :: Monad m
-                  => (PeerSharingAmount -> m [peer])
+peerSharingServer :: ( MonadSTM m
+                     , MonadMonotonicTime m
+                     , Hashable peer
+                     , RandomGen s
+                     )
+                  => PeerSharingAPI peer s m
                   -> PeerSharingServer peer m
-peerSharingServer computePeersToShare =
+peerSharingServer peerSharingAPI =
   PeerSharingServer
-    { recvMsgShareRequest = \amount -> (,) <$> computePeersToShare amount
-                                          <*> return (peerSharingServer computePeersToShare)
+    { recvMsgShareRequest = \amount -> (,) <$> computePeerSharingPeers peerSharingAPI amount
+                                           <*> return (peerSharingServer peerSharingAPI)
     }
+
+--
+-- Utility Function
+--
+
+-- | PeerSharingAPI needed to compute the peers to be shared.
+--
+data PeerSharingAPI addr s m =
+  PeerSharingAPI { psPublicPeerSelectionStateVar :: StrictTVar m (PublicPeerSelectionState addr)
+                 , psGenVar                      :: StrictTVar m s
+                 , psReSaltAtVar                 :: StrictTVar m Time
+                 , psPolicyPeerShareStickyTime   :: !DiffTime
+                 -- ^ Amount of time between changes to the salt used to pick peers to
+                 -- gossip about.
+                 , psPolicyPeerShareMaxPeers     :: !PeerSharingAmount
+                 -- ^ Maximum number of peers to respond with in a single request
+                 }
+
+-- | Amount of time between changes to the salt used to pick peers to
+-- gossip about.
+--
+ps_POLICY_PEER_SHARE_STICKY_TIME :: DiffTime
+ps_POLICY_PEER_SHARE_STICKY_TIME = 823  -- seconds
+
+-- | Maximum number of peers to respond with in a single request
+--
+ps_POLICY_PEER_SHARE_MAX_PEERS :: PeerSharingAmount
+ps_POLICY_PEER_SHARE_MAX_PEERS = 10
+
+-- | Create a new PeerSharingAPI
+--
+newPeerSharingAPI :: ( MonadSTM m
+                     , Ord addr
+                     )
+                  => s
+                  -> DiffTime
+                  -> PeerSharingAmount
+                  -> m (PeerSharingAPI addr s m)
+newPeerSharingAPI rng policyPeerShareStickyTime policyPeerShareMaxPeers = do
+  publicPeerSelectionStateVar <- newTVarIO emptyPublicPeerSelectionState
+  genVar <- newTVarIO rng
+  reSaltAtVar <- newTVarIO (Time 0)
+  return $
+    PeerSharingAPI publicPeerSelectionStateVar
+                   genVar
+                   reSaltAtVar
+                   policyPeerShareStickyTime
+                   policyPeerShareMaxPeers
+
+-- | Select a random subset of the known peers that are available to publish through peersharing.
+-- The list of peers will change after `policyPeerShareStickyTime` seconds.
+-- The list of peers shared does at most change by the number of peers added or removed.
+-- That is a newly added or removed peer can at most lead to one new peer beeing shared.
+--
+computePeerSharingPeers :: (  MonadSTM m
+                           , MonadMonotonicTime m
+                           , Hashable ntnAddr
+                           , RandomGen s
+                           )
+                        => PeerSharingAPI ntnAddr s m
+                        -> PeerSharingAmount
+                        -> m [ntnAddr]
+computePeerSharingPeers PeerSharingAPI{ psPublicPeerSelectionStateVar
+                                      , psPolicyPeerShareStickyTime
+                                      , psPolicyPeerShareMaxPeers
+                                      , psReSaltAtVar
+                                      , psGenVar
+                                      } amount = do
+  now <- getMonotonicTime
+  publicState <- readTVarIO psPublicPeerSelectionStateVar
+  salt <- atomically $ do
+    reSaltAt <- readTVar psReSaltAtVar
+    if reSaltAt <= now
+       then do
+         writeTVar psReSaltAtVar $ addTime psPolicyPeerShareStickyTime now
+         stateTVar psGenVar random
+       else do
+         gen <- readTVar psGenVar
+         return $ fst $ random gen
+
+  let availableToShareSet = availableToShare publicState
+      randomList = take (fromIntegral psPolicyPeerShareMaxPeers
+                                      `min`
+                                      fromIntegral amount)
+                 $ sortBy (\a b -> compare (hashWithSalt salt a) (hashWithSalt salt b))
+                 $ Set.elems availableToShareSet
+  if null availableToShareSet
+     then return []
+     else return randomList